+++ src/Storage/Storage/Common/AzureStorageBlob.cs
@@ -22,6 +22,13 @@ namespace Microsoft.WindowsAzure.Commands.Common.Storage.ResourceModel
     using Microsoft.Azure.Commands.Common.Authentication.Abstractions;
     using Microsoft.WindowsAzure.Commands.Storage.Common;
     using global::Azure.Storage;
+    using global::Azure.Storage.Blobs.Models;
+    using System.Collections;
+    using System.Collections.Generic;
+    using global::Azure.Storage.Blobs.Specialized;
+    using Microsoft.Azure.Storage.Auth;
+    using Microsoft.Azure.Storage;
+    using System.Threading.Tasks;
 
     /// <summary>
     /// Azure storage blob object
@@ -40,7 +47,7 @@ namespace Microsoft.WindowsAzure.Commands.Common.Storage.ResourceModel
         /// Azure storage blob type
         /// </summary>
         [Ps1Xml(Label = "BlobType", Target = ViewControl.Table, Position = 1, TableColumnWidth = 9)]
-        public BlobType BlobType { get; private set; }
+        public Microsoft.Azure.Storage.Blob.BlobType BlobType { get; private set; }
 
         /// <summary>
         /// Blob length
@@ -55,7 +62,7 @@ namespace Microsoft.WindowsAzure.Commands.Common.Storage.ResourceModel
         public bool IsDeleted { get; private set; }
 
         /// <summary>
-        /// XSCL Track2 Blob Client, used to run blob APIs
+        /// XSCL Track2 Blob Client, used to run blob APIs, this property should be removed in the future major release, and replace with BlobBaseClient
         /// </summary>
         public BlobClient BlobClient
         {
@@ -63,12 +70,48 @@ namespace Microsoft.WindowsAzure.Commands.Common.Storage.ResourceModel
             {
                 if (privateBlobClient == null)
                 {
-                    privateBlobClient = GetTrack2BlobClient(this.ICloudBlob, (AzureStorageContext)this.Context);
+                    if (this.ICloudBlob != null)
+                    {
+                        privateBlobClient = GetTrack2BlobClient(this.ICloudBlob, (AzureStorageContext)this.Context, this.privateClientOptions);
+                    }
+                    else if(this.privateBlobBaseClient != null)
+                    {
+                        privateBlobClient = GetTrack2BlobClient(this.privateBlobBaseClient, (AzureStorageContext)this.Context, this.privateClientOptions);
+                    }
                 }
                 return privateBlobClient;
             }
         }
         private BlobClient privateBlobClient = null;
+
+        /// <summary>
+        /// XSCL Track2 Blob Client, used to run blob APIs
+        /// </summary>
+        public BlobBaseClient BlobBaseClient
+        {
+            get
+            {
+                if (privateBlobBaseClient == null)
+                {
+                    if (this.privateBlobClient == null && this.ICloudBlob != null && !(this.ICloudBlob is InvalidCloudBlob))
+                    {
+                        privateBlobClient = GetTrack2BlobClient(this.ICloudBlob, (AzureStorageContext)this.Context, this.privateClientOptions);
+                    }
+                    var Track2BlobType = Util.convertBlobType_Track1ToTrack2(this.BlobType);
+                    if (Track2BlobType != null)
+                    {
+                        privateBlobBaseClient = Util.GetTrack2BlobClientWithType(privateBlobClient, (AzureStorageContext)this.Context, Track2BlobType.Value, this.privateClientOptions);
+                    }
+                    else
+                    {
+                        privateBlobBaseClient = (BlobBaseClient)privateBlobClient;
+                    }
+                    // Don't need add versionID, since with VersionID the object must be get from Track2SDK, and the privateBlobBaseClient should no be null in this case
+                }
+                return privateBlobBaseClient;
+            }
+        }
+        private BlobBaseClient privateBlobBaseClient = null;
 
         /// <summary>
         /// XSCL Track2 Blob properties, will retrieve the properties on server and return to user
@@ -79,7 +122,7 @@ namespace Microsoft.WindowsAzure.Commands.Common.Storage.ResourceModel
             {
                 if (privateBlobProperties == null)
                 {
-                    privateBlobProperties = BlobClient.GetProperties().Value;
+                    privateBlobProperties = BlobBaseClient.GetProperties().Value;
                 }
                 return privateBlobProperties;
             }
@@ -116,10 +159,37 @@ namespace Microsoft.WindowsAzure.Commands.Common.Storage.ResourceModel
         public BlobContinuationToken ContinuationToken { get; set; }
 
         /// <summary>
+        /// Blob VersionId.
+        /// </summary>
+        public string VersionId { get; set; }
+
+        /// <summary>
+        /// Blob IsCurrentVersion..
+        /// </summary>
+        public bool? IsLatestVersion { get; set; }
+
+        private BlobClientOptions privateClientOptions = null;
+
+        /// <summary>
+        /// Blob AccessTier..
+        /// </summary>
+        public string AccessTier { get; set; }
+
+        /// <summary>
+        /// Blob TagCount.
+        /// </summary>
+        public long TagCount { get; set; }
+
+        /// <summary>
+        /// Blob Tags
+        /// </summary>
+        public Hashtable Tags { get; set; }
+
+        /// <summary>
         /// Azure storage blob constructor
         /// </summary>
         /// <param name="blob">ICloud blob object</param>
-        public AzureStorageBlob(CloudBlob blob, AzureStorageContext storageContext)
+        public AzureStorageBlob(CloudBlob blob, AzureStorageContext storageContext, BlobClientOptions options = null)
         {
             Name = blob.Name;
             ICloudBlob = blob;
@@ -131,16 +201,163 @@ namespace Microsoft.WindowsAzure.Commands.Common.Storage.ResourceModel
             LastModified = blob.Properties.LastModified;
             SnapshotTime = blob.SnapshotTime;
             this.Context = storageContext;
+            this.privateClientOptions = options;
+            AccessTier = blob.Properties.StandardBlobTier is null ?
+                (blob.Properties.PremiumPageBlobTier is null ? null : blob.Properties.PremiumPageBlobTier.ToString())
+                : blob.Properties.StandardBlobTier.ToString();
+        }
+
+        /// <summary>
+        /// Azure storage blob constructor
+        /// </summary>
+        /// <param name="blob">ICloud blob object</param>
+        public AzureStorageBlob(BlobBaseClient track2BlobClient, AzureStorageContext storageContext, BlobClientOptions options = null, BlobItem listBlobItem = null)
+        {
+            if (listBlobItem == null)
+            {
+                SetProperties(track2BlobClient, storageContext, track2BlobClient.GetProperties().Value, options);
+                return;
+            }
+
+            this.privateBlobBaseClient = track2BlobClient;
+            Name = track2BlobClient.Name;
+            this.Context = storageContext;
+            privateClientOptions = options;
+            ICloudBlob = GetTrack1Blob(track2BlobClient, storageContext.StorageAccount.Credentials, listBlobItem.Properties.BlobType);
+            if (!(ICloudBlob is InvalidCloudBlob))
+            {
+                BlobType = ICloudBlob.BlobType;
+                SnapshotTime = ICloudBlob.SnapshotTime;
+            }
+            else
+            {
+                BlobType = Util.convertBlobType_Track2ToTrack1(listBlobItem.Properties.BlobType);
+                if (listBlobItem.Snapshot != null)
+                {
+                    SnapshotTime = DateTimeOffset.Parse(listBlobItem.Snapshot);
+                }
+            }
+
+            // Set the AzureStorageBlob Properties
+            Length = listBlobItem.Properties.ContentLength is null ? 0 : listBlobItem.Properties.ContentLength.Value;
+            IsDeleted = listBlobItem.Deleted;
+            RemainingDaysBeforePermanentDelete = listBlobItem.Properties.RemainingRetentionDays;
+            ContentType = listBlobItem.Properties.ContentType;
+            LastModified = listBlobItem.Properties.LastModified;
+            VersionId = listBlobItem.VersionId;
+            IsLatestVersion = listBlobItem.IsLatestVersion;
+            AccessTier = listBlobItem.Properties.AccessTier is null? null: listBlobItem.Properties.AccessTier.ToString();
+            if (listBlobItem.Tags != null)
+            {
+                Tags = listBlobItem.Tags.ToHashtable();
+                TagCount = listBlobItem.Tags.Count;
+            }
+        }
+
+
+        public AzureStorageBlob(BlobBaseClient track2BlobClient, AzureStorageContext storageContext, global::Azure.Storage.Blobs.Models.BlobProperties blobProperties, BlobClientOptions options = null)
+        {
+            SetProperties(track2BlobClient, storageContext, blobProperties, options);
+        }
+
+        private void SetProperties(BlobBaseClient track2BlobClient, AzureStorageContext storageContext, global::Azure.Storage.Blobs.Models.BlobProperties blobProperties = null, BlobClientOptions options = null)
+        {
+            if (blobProperties == null)
+            {
+                try
+                {
+                    privateBlobProperties = track2BlobClient.GetProperties().Value;
+                }
+                catch (global::Azure.RequestFailedException e) when (e.Status == 403 || e.Status == 404)
+                {
+                    // privateBlobProperties will be null when there are no permission to get blob proeprties, or blob is already deleted.
+                }
+            }
+            else
+            {
+                privateBlobProperties = blobProperties;
+            }
+
+            this.privateBlobBaseClient = track2BlobClient;
+            Name = track2BlobClient.Name;
+            this.Context = storageContext;
+            privateClientOptions = options;
+            ICloudBlob = GetTrack1Blob(track2BlobClient, storageContext.StorageAccount.Credentials, privateBlobProperties.BlobType);
+            if (!(ICloudBlob is InvalidCloudBlob))
+            {
+                BlobType = ICloudBlob.BlobType;
+                SnapshotTime = ICloudBlob.SnapshotTime;
+            }
+            else // This code might should not be necessary, since currently only blob version will has Track1 Blob as null, and blob veresion won't have snapshot time
+            {
+                SnapshotTime = Util.GetSnapshotTimeFromBlobUri(track2BlobClient.Uri);
+            }
+
+            // Set the AzureStorageBlob Properties
+            if (privateBlobProperties != null)
+            {
+                Length = privateBlobProperties.ContentLength;
+                ContentType = privateBlobProperties.ContentType;
+                LastModified = privateBlobProperties.LastModified;
+                VersionId = privateBlobProperties.VersionId;
+                IsLatestVersion = privateBlobProperties.IsLatestVersion;
+                if (ICloudBlob is InvalidCloudBlob)
+                {
+                    BlobType = Util.convertBlobType_Track2ToTrack1(privateBlobProperties.BlobType);
+                }
+                AccessTier = privateBlobProperties.AccessTier is null ? null : privateBlobProperties.AccessTier.ToString();
+                TagCount = privateBlobProperties.TagCount;
+            }
+        }
+
+        /// <summary>
+        /// Get Track1 Blob Object
+        /// Will return null if it's a Blob version, since Track1 not support blob version
+        /// </summary>
+        /// <param name="track2BlobClient"></param>
+        public static CloudBlob GetTrack1Blob(BlobBaseClient track2BlobClient, StorageCredentials credentials, global::Azure.Storage.Blobs.Models.BlobType? blobType = null)
+        {
+            if (Util.GetVersionIdFromBlobUri(track2BlobClient.Uri) != null)
+            {
+                // Track1 SDK don't support blob VersionId
+                return new InvalidCloudBlob(track2BlobClient.Uri, credentials);
+            }
+
+            if (credentials.IsSAS) // the Uri already contains credentail.
+            {
+                credentials = null;
+            }
+            CloudBlob track1Blob;
+            if (blobType == null)
+            {
+                track1Blob = new CloudBlob(track2BlobClient.Uri, credentials);
+            }
+            else
+            {
+                switch (blobType.Value)
+                {
+                    case global::Azure.Storage.Blobs.Models.BlobType.Page:
+                        track1Blob = new CloudPageBlob(track2BlobClient.Uri, credentials);
+                        break;
+                    case global::Azure.Storage.Blobs.Models.BlobType.Append:
+                        track1Blob = new CloudAppendBlob(track2BlobClient.Uri, credentials);
+                        break;
+                    default: //Block
+                        track1Blob = new CloudBlockBlob(track2BlobClient.Uri, credentials);
+                        break;
+                }
+            }
+            return track1Blob;
         }
 
         //refresh XSCL track2 blob properties object from server
         public void FetchAttributes()
         {
-            privateBlobProperties = BlobClient.GetProperties().Value;
+            privateBlobProperties = BlobBaseClient.GetProperties().Value;
         }
 
         // Convert Track1 Blob object to Track 2 blob Client
-        protected static BlobClient GetTrack2BlobClient(CloudBlob cloubBlob, AzureStorageContext context)
+        public static BlobClient GetTrack2BlobClient(CloudBlob cloubBlob, AzureStorageContext context, BlobClientOptions options = null)
         {
             BlobClient blobClient;
             if (cloubBlob.ServiceClient.Credentials.IsToken) //Oauth
@@ -150,7 +367,7 @@ namespace Microsoft.WindowsAzure.Commands.Common.Storage.ResourceModel
                     //TODO : Get Oauth context from current login user.
                     throw new System.Exception("Need Storage Context to convert Track1 Blob object in token credentail to Track2 Blob object.");
                 }
-                blobClient = new BlobClient(cloubBlob.SnapshotQualifiedUri, context.Track2OauthToken);
+                blobClient = new BlobClient(cloubBlob.SnapshotQualifiedUri, context.Track2OauthToken, options);
 
             }
             else if (cloubBlob.ServiceClient.Credentials.IsSAS) //SAS
@@ -165,16 +382,42 @@ namespace Microsoft.WindowsAzure.Commands.Common.Storage.ResourceModel
                 {
                     fullUri = fullUri + cloubBlob.ServiceClient.Credentials.SASToken;
                 }
-                blobClient = new BlobClient(new Uri(fullUri));
+                blobClient = new BlobClient(new Uri(fullUri), options);
             }
             else if (cloubBlob.ServiceClient.Credentials.IsSharedKey) //Shared Key
             {
                 blobClient = new BlobClient(cloubBlob.SnapshotQualifiedUri,
-                    new StorageSharedKeyCredential(context.StorageAccountName, cloubBlob.ServiceClient.Credentials.ExportBase64EncodedKey()));
+                    new StorageSharedKeyCredential(context.StorageAccountName, cloubBlob.ServiceClient.Credentials.ExportBase64EncodedKey()), options);
             }
             else //Anonymous
             {
-                blobClient = new BlobClient(cloubBlob.SnapshotQualifiedUri);
+                blobClient = new BlobClient(cloubBlob.SnapshotQualifiedUri, options);
+            }
+
+            return blobClient;
+        }
+
+        // Convert Blob object to Track 2 blob Client
+        public static BlobClient GetTrack2BlobClient(BlobBaseClient blobBaseClient, AzureStorageContext context, BlobClientOptions options = null)
+        {
+            if (blobBaseClient is BlobClient)
+            {
+                return (BlobClient)blobBaseClient;
+            }
+            BlobClient blobClient;
+            if (context.StorageAccount.Credentials.IsToken) //Oauth
+            {
+                blobClient = new BlobClient(blobBaseClient.Uri, context.Track2OauthToken, options);
+
+            }
+            else if (context.StorageAccount.Credentials.IsSharedKey) //Shared Key
+            {
+                blobClient = new BlobClient(blobBaseClient.Uri,
+                    new StorageSharedKeyCredential(context.StorageAccountName, context.StorageAccount.Credentials.ExportBase64EncodedKey()), options);
+            }
+            else //Anonymous or SAS
+            {
+                blobClient = new BlobClient(blobBaseClient.Uri, options);
             }
 
             return blobClient;